+++ Midi/Sanford.Multimedia.Midi/Device Classes/OutputDevice Classes/OutputDevice.cs
@@ -48,11 +48,11 @@ namespace Sanford.Multimedia.Midi
         #region Win32 Midi Output Functions and Constants
 
         [DllImport("winmm.dll")]
-        private static extern int midiOutOpen(ref int handle, int deviceID,
+        private static extern int midiOutOpen(ref IntPtr handle, int deviceID,
             MidiOutProc proc, int instance, int flags);
 
         [DllImport("winmm.dll")]
-        private static extern int midiOutClose(int handle);
+        private static extern int midiOutClose(IntPtr handle);
 
         #endregion
